Renamed pidROS to pid_ros
